Update patch set 4
Patch Set 4:

Hi Katri,

I see that you have env variables for host and port of ves collector.
However, in order to push to VES collector, ves agent needs to use the correct URL, which also requires path; this should be configurable.  Here is the URL it would use:
http://ricaux-entry/vescollector/eventListener/v7

ricaux-entry is the hostname, which refers to a kubernetes service; no port is given.
/vescollector is used as the first path element so that kong ingress controller will send to VES
/eventListener/v7 is the URL path which VES collector will actually receive.  Please note that if VES v5 is used it should be v5.

I believe that all of these URL elements should be configurable

Please see for more info:
https://rancodev.atlassian.net/wiki/spaces/IT/pages/100958213/RIC+Measurement+VES+Flow

Patch-set: 4