Configure xApp metrics in xApp description
Vesmgr subscribes the xApp notifications from xAppMgr. When it receives a
notification, it requests the xApp config from xAppMgr, creates the VESPA
configuration according to it, and restarts VESPA with the new configuration.

The xApp counters should be defined in the xApp dcescriptor. Vesmgr reads the
counter defitions from section config->metrics. The following fields
are required:

* name - Prometheus name of the counter
* objectName - Ves collector object name
* objectInstance - Ves collector object instance

If the same counter name is defined by several xApps, vesmgr makes the
configuration based on the first definition and ignores the latter ones,
if they are conflicting. However, VESPA reports all counters to Ves collector
regarless of the xApp exposing them.

VESPA reads the ricComponentName from Prometheus label "kubernetes_name".

There are no more hard-coded metrics configured in this version.

Change-Id: I746c6941cebe686165aed97e223b0ec0e9c7a679
Signed-off-by: Katri Turunen <katri.turunen@nokia.com>